fix: rename errors to exceptions
